последние штрихи

+++ app/Http/Controllers/CartController.php
@@ -7,18 +7,23 @@ use App\Models\Cart;
 use App\Models\Product;
 
 use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Validation\ValidationException;
+
 
 class CartController extends Controller
 {
-    public function showCart(Authenticatable $user)
+    public function showCart(Request $request)
     {
         try {
+            $user = $request->user();
+
+            if (!$user) {
+                return response()->json(['error' => 'Пользователь не найден'], 404);
+            }
+
             $cartItems = $user->cart()->with('product')->get();
 
             return response()->json(['cart_items' => $cartItems]);
@@ -28,40 +33,56 @@ class CartController extends Controller
     }
 
 
-    public function addToCart($product_id,$quantity)
+    public function addToCart(Request $request, $product_id, $quantity)
     {
-        $user_id=Auth::id();
+        try {
+            $user = $request->user();
 
+            if (!$user) {
+                return response()->json(['error' => 'Пользователь не найден'], 404);
+            }
 
-        $product = Product::find ($product_id);
+            $product = Product::find($product_id);
 
-        if (!$product){
-            return response()->json(['message' => 'товар не найден']);
+            if (!$product) {
+                return response()->json(['error' => 'Товар не найден'], 404);
+            }
+
+            // Получаем текущее количество товара в корзине для конкретного продукта пользователя
+            $currentTotalItems = Cart::where('user_id', $user->id)
+                ->where('product_id', $product_id)
+                ->value('total_items');
+
+            $newTotalItems = $currentTotalItems + $quantity;
+
+            if ($newTotalItems > $product->stock_quantity) {
+                return response()->json(['error' => 'Количество товара в корзине не может превышать общее количество на складе', 'stock_quantity' => $product->stock_quantity], 400);
+            }
+
+            $cartItem = Cart::where('user_id', $user->id)
+                ->where('product_id', $product_id)
+                ->first();
+
+            if ($cartItem) {
+                // Обновляем существующий элемент корзины
+                $cartItem->total_items += $quantity;
+                $cartItem->total_price += ($product->price * $quantity);
+                $cartItem->save();
+            } else {
+                // Создаем новый элемент корзины
+                Cart::create([
+                    'user_id' => $user->id,
+                    'product_id' => $product_id,
+                    'total_items' => $quantity,
+                    'total_price' => ($product->price * $quantity),
+                ]);
+            }
+
+            return response()->json(['message' => 'Товар успешно добавлен в корзину', 'stock_quantity' => $product->stock_quantity]);
+        } catch (\Exception $e) {
+            return response()->json(['error' => 'Не удалось добавить товар в корзину', 'details' => $e->getMessage()], 500);
         }
-        $cartItem = Cart::where('user_id', $user_id)
-            ->where('product_id', $product_id)
-            ->first();
-        if ($cartItem) {
-
-            $cartItem->total_items += $quantity;
-            $cartItem->total_price += ($product->price * $quantity);
-            $cartItem->save();
-        } else {
-
-            Cart::create([
-                'user_id' => $user_id,
-                'product_id' => $product_id,
-                'total_items' => $quantity,
-                'total_price' => ($product->price * $quantity),
-
-            ]);
-        }
-
-        return response()->json(['message' => 'Товар успешно добавлен в корзину']);
-
     }
-
-
     public function updateProductFromCart(CartRequest $request, $product_id)
     {
         try {
@@ -74,18 +95,24 @@ class CartController extends Controller
 
             try {
                 $total_item = $request->input('total_item');
-                $product=$cartItem->product;
-                $cartItem->total_items = $total_item;
-                $cartItem->total_price = ($product->price * $total_item);
-                $cartItem->save();
+                $product = $cartItem->product;
 
-                DB::commit();
+                // Проверка, чтобы не превышать общее количество товара в корзине
+                if ($total_item <= $product->stock_quantity) {
+                    $cartItem->total_items = $total_item;
+                    $cartItem->total_price = ($product->price * $total_item);
+                    $cartItem->save();
+                    DB::commit();
+
+                    return response()->json(['message' => 'Товар в корзине обновлен успешно'], 200);
+                } else {
+                    DB::rollBack();
+                    return response()->json(['message' => 'Количество товара в корзине не может превышать общее количество на складе'], 400);
+                }
             } catch (\Exception $e) {
                 DB::rollBack();
                 throw $e;
             }
-
-            return response()->json(['message' => 'Товар в корзине обновлен успешно'], 200);
         } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Товар не найден в корзине'], 404);
         } catch (AuthorizationException $e) {
@@ -102,7 +129,7 @@ class CartController extends Controller
             $cartItem = Cart::where('product_id', $product_id)->firstOrFail();
 
 
-//            $this->authorize('delete', $cartItem);
+            $this->authorize('delete', $cartItem);
 
 
             $cartItem->delete();
@@ -119,16 +146,5 @@ class CartController extends Controller
         }
     }
 
-    public function clear()
-    {
-        try {
-
-            Cart::truncate();
-
-            return response()->json(['message' => 'Корзина очищена'], 200);
-        } catch (\Exception $e) {
-            return response()->json(['message' => 'Что-то пошло не так'], 500);
-        }
-    }
 
 }
